add: added argument --keep-junk
Now is automatically deleted files in tmp and wav track on force close or end. You can keep them adding --keep-junk

+++ simple_ehm-runnable.py
@@ -6,7 +6,7 @@ import os
 import io
 import pathlib
 import argparse
-
+import signal
 import matplotlib.pyplot as plt
 import numpy as np
 import seaborn as sns
@@ -41,6 +41,7 @@ parser.add_argument("--generate-training-data", default=False, action='store_tru
 parser.add_argument("--srt", default=False, action='store_true', help="generate subtitle track for easier accuracy evaluation")
 parser.add_argument("--keep", nargs="+", default=["speech"], help="space separated tags to to be kept in the final video. Eg: ehm silence. Default: speech")
 parser.add_argument("--name", nargs="+", default=False, help="Output video name")
+parser.add_argument("--keep-junk", default=False,action="store_true", help="keeps tmp files")
 
 args = parser.parse_args()
 video_path = args.filename
@@ -56,6 +57,25 @@ trash = set()
 
 cuts = []       # edits for ffmpeg to split
 mergelist = []  # files for ffmpeg to merge back
+
+# ctrl+c handler
+def signal_handler(sig, frame):
+    try:
+        filename, file_extension = os.path.splitext(video_path)
+        
+        if(not args.keep_junk):
+            os.remove(f'{filename}.wav')
+            for file in os.listdir("tmp"):
+                os.remove("tmp/"+file)
+            #print(f"file: {file} removed")
+
+    except Exception as e:
+        #print(str(e))
+        pass
+
+    exit()
+
+signal.signal(signal.SIGINT, signal_handler)
 
 def plot_spectrogram(spectrogram, ax):
   # Convert to frequencies to log scale and transpose so that the time is
@@ -316,7 +336,11 @@ if __name__ == '__main__':
     analyze_track(model, waveform, sample_rate)
     cut_and_merge(video_path[:-4])
     
-    os.remove(wav_path)
+	# removes tmp files
+    if(not args.keep_junk):
+        os.remove(wav_path)
+        for file in os.listdir("tmp"):
+            os.remove("tmp/"+file)
 
     print("\nFATTO!")
     for k, v in zip(_perf.keys(), _perf.values()):
